+++ NEWS.md
@@ -7,6 +7,8 @@ None
 1. Fix #405 - Add function `quantile_normalization()` to normalize data using quantiles.
 2. Fix #409 - Add function `check_duplilcate_rows()` to check for duplicate rows in a data frame.
 3. Fix #414 - Add function `util_chisquare_param_estimate()` to estimate the parameters of the chi-square distribution.
+4. Fix #417 - Add function `tidy_mcmc_sampling()` to sample from a distribution using MCMC.
+This outputs the function sampled data and a diagnostic plot.
 
 ## Minor Fixes and Improvements
 1. Fix #401 - Update `tidy_multi_single_dist()` to respect the `.return_tibble` parameter